+++ CHANGELOG.md
@@ -10,3 +10,24 @@ Log of unreleased changes.
 
 - CHANGELOG.md
 - You will be able to open a drop-down with a closed parent.
+- Implemented custom events `open`, `close`, `openAll`, `closeAll`, `init` and
+`destroy`.
+- `preventDefault` set to `true` will let you open the link in a new tab using
+`Ctrl + Click`.
+
+### Changed
+
+- Updated `README.md` adding *Events* section.
+
+### Deprecated
+
+- Callbacks are deprecated in favor of events in `v2.0.0`.
+- Support for IE9 and IE10 will be removed in `v2.0.0`
+
+### Removed
+
+- Removed `.eslintrc.json` file.
+
+## v1.2.0
+
+...